requirements.txt
appdirs==1.4.3
Django==1.11
djangorestframework==3.6.2
geojson==1.3.5
numpy==1.12.1
packaging==16.8
pandas==0.19.2
pymongo==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
scikit-learn==0.18.1
scipy==0.19.0
six==1.10.0
sklearn==0.0
